+++ spare/tests/conftest.py
@@ -5,11 +5,10 @@ import secrets
 import sys
 import tempfile
 
-from botocore.client import Config
-from boto3 import resource
 from mirakuru import HTTPExecutor
 from pathlib import Path
 from spare.envoy import Envoy
+from spare.utils import s3_client
 from urllib.request import urlretrieve
 
 
@@ -40,19 +39,14 @@ def s3(access_key, secret_key, endpoint, minio_path):
     # minio has some concurrency problems with locks which compound if we
     # use the default connection setup which includes lots of retries and
     # big timeouts
-    config = Config(
+    s3 = s3_client(
+        endpoint.geturl(),
+        access_key,
+        secret_key,
         connect_timeout=1,
         read_timeout=1,
-        retries={'max_attempts': 3}
+        max_attempts=3
     )
-
-    s3 = resource(
-        service_name='s3',
-        use_ssl=False,
-        endpoint_url=endpoint.geturl(),
-        aws_access_key_id=access_key,
-        aws_secret_access_key=secret_key,
-        config=config)
 
     yield s3